replaced libft with newest version, qnd branch to get points faster

+++ libft/srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: olkovale <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/07 11:29:19 by olkovale          #+#    #+#             */
-/*   Updated: 2017/06/21 13:12:09 by olkovale         ###   ########.fr       */
+/*   Updated: 2017/09/05 19:30:21 by olkovale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int		ft_strlen(const char *str)
 {
-	size_t i;
+	int		i;
 
 	i = 0;
 	while (str[i])
