+++ libft/srcs/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: olkovale <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/11 02:25:54 by olkovale          #+#    #+#             */
-/*   Updated: 2017/08/22 10:12:45 by olkovale         ###   ########.fr       */
+/*   Updated: 2017/09/06 15:12:24 by olkovale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 #include "libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, size_t sz)
+int		ft_strncmp(const char *s1, const char *s2, int sz)
 {
 	unsigned char	c1;
 	unsigned char	c2;
